scard: add APDU I/O interface support for T=0 protocol
Sun/Oracle's JCK uses TLP224 messages to talk with JCWDE if protocol is
set to T=0. To support developing JavaCard applets, add (limited) support
for TLP224 protocol in form of apduio_t0 transport.

Note: cref from JCK is limited to T=1 only, so it can not be used with scard
library for now.

Signed-off-by: Dmitry Eremin-Solenikov <dbaryshkov@gmail.com>